调整小问题

+++ src/main/java/com/example/demo/dto/MotorFanDTO.java
@@ -6,7 +6,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 /**
@@ -27,14 +26,14 @@ public class MotorFanDTO {
     private Long probeSensorId;          // 探头传感器ID
     
     // 温控/循环共用字段
-    private BigDecimal tempUpper;        // 温度上限/启动温度（°C）
-    private BigDecimal tempLower;        // 温度下限/停止温度（°C）
+    private Double tempUpper;            // 温度上限/启动温度（°C）
+    private Double tempLower;            // 温度下限/停止温度（°C）
     private Integer runTime;             // 运行时间/低温运行（秒）
     private Integer pauseTime;           // 暂停时间/低温暂停（秒）
     
     // 湿控字段
-    private BigDecimal humidityUpper;    // 湿度上限（%）
-    private BigDecimal humidityLower;    // 湿度下限（%）
+    private Double humidityUpper;        // 湿度上限（%）
+    private Double humidityLower;        // 湿度下限（%）
     
     // 气体字段
     private Integer gasUpper;            // 气体上限（ppm）
@@ -45,24 +44,24 @@ public class MotorFanDTO {
     private String timer1StartTime;      // 定时1开启时间（HH:mm）
     private String timer1EndTime;        // 定时1关闭时间（HH:mm）
     private Long timer1ProbeSensorId;    // 定时1探头传感器ID
-    private BigDecimal timer1StartTemp;  // 定时1启动温度（°C）
-    private BigDecimal timer1StopTemp;   // 定时1停止温度（°C）
+    private Double timer1StartTemp;      // 定时1启动温度（°C）
+    private Double timer1StopTemp;       // 定时1停止温度（°C）
     
     // 定时2字段
     private Integer timer2Enabled;       // 定时2开关：0-关闭 1-开启
     private String timer2StartTime;      // 定时2开启时间（HH:mm）
     private String timer2EndTime;        // 定时2关闭时间（HH:mm）
     private Long timer2ProbeSensorId;    // 定时2探头传感器ID
-    private BigDecimal timer2StartTemp;  // 宝时2启动温度（°C）
-    private BigDecimal timer2StopTemp;   // 定时2停止温度（°C）
+    private Double timer2StartTemp;      // 定时2启动温度（°C）
+    private Double timer2StopTemp;       // 定时2停止温度（°C）
     
     // 定时3字段
     private Integer timer3Enabled;       // 定时3开关：0-关闭 1-开启
     private String timer3StartTime;      // 定时3开启时间（HH:mm）
     private String timer3EndTime;        // 定时3关闭时间（HH:mm）
     private Long timer3ProbeSensorId;    // 定时3探头传感器ID
-    private BigDecimal timer3StartTemp;  // 定时3启动温度（°C）
-    private BigDecimal timer3StopTemp;   // 定时3停止温度（°C）
+    private Double timer3StartTemp;      // 定时3启动温度（°C）
+    private Double timer3StopTemp;       // 定时3停止温度（°C）
     
     @JsonFormat(pattern = DateUtils.DATE_TIME_FORMAT, timezone = DateUtils.TIME_ZONE)
     private LocalDateTime createdTime;    // 创建时间
